requirements.txt
aiohttp==3.8.1
APScheduler
async-timeout
chardet
discord.py==1.7.3
emoji
pytz
six
tzlocal==2.0
websockets
requests
htmldom
wikipedia
BeautifulSoup4
matplotlib
Pillow
discord_config
mee6-py-api
python-dotenv
pytube
requests-html
python-dateutil
TenGiphPy
py_expression_eval
prettytable
windyLib
